refactor(mcp): split tool_answer.py into a tool_answer/ package
Break the 1356-line get_answer module into an orchestrator + focused
sibling modules behind a package façade. `from ...tool_answer import
get_answer` resolves unchanged and importing the package still registers
the tool via @mcp.tool().

- config.py: all tuning constants + the system/user prompt templates.
- symbols.py: question-identifier extraction + WikiSymbol hydration
  (_read_symbol_source / _read_signature_from_source).
- retrieval.py: coverage rerank, domain penalty, intersection boost,
  gated-excerpt enrichment, candidate justification.
- synthesis.py: provider resolution + question hashing.
- confidence.py: the hedge-detection predicate.
- answer.py: get_answer orchestrator only.

Verbatim moves; no logic changes. Dropped the dead v1 `_build_context_block`
(zero callers — superseded by _answer_context.build_context_block). Pre-existing
ambiguous-unicode (en-dash/×) in prompts kept verbatim to preserve LLM input.
tests/unit/server (259, incl. 42 test_mcp) green; ruff clean on new modules;
no new mypy errors.
